potato_leaf_disease_detection/requirements.txt
Pillow
numpy
requests
streamlit
tensorflow
